Endret struktur på opprettelse av kravgrunnlag, statsmelding og tilbakekreving
