+++ autotest/src/test/kotlin/no/nav/familie/tilbake/e2e/autotest/OpprettTilbakekrevingOS.kt
@@ -13,8 +13,8 @@ import no.nav.familie.tilbake.e2e.domene.dto.SærligGrunn
 import no.nav.familie.tilbake.e2e.domene.dto.Venteårsak
 import no.nav.familie.tilbake.e2e.domene.dto.Vilkårsvurderingsresultat
 import no.nav.familie.tilbake.e2e.klient.FamilieTilbakeKlient
-import no.nav.familie.tilbake.e2e.klient.OpprettKravgrunnlagBuilder
-import no.nav.familie.tilbake.e2e.klient.OpprettTilbakekrevingBuilder
+import no.nav.familie.tilbake.e2e.domene.builder.OpprettKravgrunnlagBuilder
+import no.nav.familie.tilbake.e2e.domene.builder.OpprettTilbakekrevingBuilder
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
@@ -35,9 +35,7 @@ class OpprettTilbakekrevingOS(@Autowired val familieTilbakeKlient: FamilieTilbak
 
     @BeforeEach
     fun setup() {
-        saksbehandler = Saksbehandler(familieTilbakeKlient = familieTilbakeKlient,
-                                      opprettTilbakekrevingBuilder = OpprettTilbakekrevingBuilder(),
-                                      opprettKravgrunnlagBuilder = OpprettKravgrunnlagBuilder())
+        saksbehandler = Saksbehandler(familieTilbakeKlient = familieTilbakeKlient)
     }
 
     @Test
